MAINT: modifications after my code review and building docs
